Use TypeId as unique arm identifier
Using `line!()` is bound to subtly fail if you write the different cases
in the same line. This can happen unnoticedly if it is used e.g. in a
macro.

This PR replaces the use of `line!()` with the `TypeId` of an ephemeral
type. That ensures that the index is actually unique under all
circumstances. `TypeId` is a wrapper around a `u64`, so there should be
no drawbacks to this approach.